+++ openlane/config/pdk.py
@@ -468,14 +468,14 @@ scl_variables = [
     ),
     Variable(
         "GPL_CELL_PADDING",
-        Optional[Decimal],
-        "Cell padding value (in sites) for global placement. Using this is not recommended as you can simply use the density control for global placement.",
+        Decimal,
+        "Cell padding value (in sites) for global placement. The number will be integer divided by 2 and placed on both sides.",
         units="sites",
     ),
     Variable(
         "DPL_CELL_PADDING",
         Decimal,
-        "Defines the number of sites to pad the cells lef views with during detailed placement. The number will be integer divided by 2 and placed on both sides.",
+        "Cell padding value (in sites) for detailed placement. The number will be integer divided by 2 and placed on both sides. Should be <= global placement.",
         units="sites",
     ),
     Variable(
@@ -488,12 +488,6 @@ scl_variables = [
         "DIODE_CELL",
         Optional[str],
         "Defines a diode cell used to fix antenna violations, in the format {name}/{port}.",
-    ),
-    Variable(
-        "HEURISTIC_ANTENNA_THRESHOLD",
-        Optional[Decimal],
-        "A manhattan distance above which a diode is recommended to be inserted by a heuristic inserter. If not specified, the heuristic inserter will typically use a default value.",
-        units="µm",
     ),
     # Routing
     Variable(