+++ openlane/steps/odb.py
@@ -15,6 +15,7 @@ import os
 import re
 import sys
 import json
+from decimal import Decimal
 from abc import abstractmethod
 from typing import List, Optional
 
@@ -26,6 +27,7 @@ from .common_variables import io_layer_variables
 
 from ..config import Path, Variable, StringEnum
 from ..common import get_openlane_root, get_script_dir
+from ..logging import warn
 
 inf_rx = re.compile(r"\b(-?)inf\b")
 
@@ -277,6 +279,54 @@ class CustomIOPlacement(OdbpyStep):
 
 
 @Step.factory.register()
+class DiodesOnPorts(OdbpyStep):
+    id = "Odb.DiodesOnPorts"
+    name = "Diodes on Ports"
+    long_name = "Diode on Port Insertion Script"
+
+    config_vars = [
+        Variable(
+            "DIODE_ON_PORTS",
+            StringEnum("DIODE_ON_PORTS", ["none", "in", "out", "both"]),
+            "Always insert diodes on ports with the specified polarities.",
+            default="none",
+        ),
+    ]
+
+    def get_script_path(self):
+        return os.path.join(get_script_dir(), "odbpy", "diodes.py")
+
+    def get_subcommand(self) -> List[str]:
+        return ["place"]
+
+    def get_command(self) -> List[str]:
+        cell, pin = self.config["DIODE_CELL"].split("/")
+
+        return super().get_command() + [
+            "--threshold",
+            "Infinity",
+            "--diode-cell",
+            cell,
+            "--diode-pin",
+            pin,
+            "--port-protect",
+            self.config["DIODE_ON_PORTS"].value,
+        ]
+
+    def run(self, **kwargs) -> State:
+        if self.config["DIODE_ON_PORTS"].value == "none":
+            warn("'DIODE_ON_PORTS' is set to 'none': skipping…")
+            return Step.run(self, **kwargs)
+
+        if self.config["GPL_CELL_PADDING"] == 0:
+            warn(
+                "'GPL_CELL_PADDING' is set to 0. This step may cause overlap failures."
+            )
+
+        return super().run(**kwargs)
+
+
+@Step.factory.register()
 class HeuristicDiodeInsertion(OdbpyStep):
     id = "Odb.HeuristicDiodeInsertion"
     name = "Heuristic Diode Insertion"
@@ -291,10 +341,10 @@ class HeuristicDiodeInsertion(OdbpyStep):
             default=False,  # For compatibility with OL1. Yep.
         ),
         Variable(
-            "DIODE_ON_PORTS",
-            StringEnum("DIODE_ON_PORTS", ["none", "in", "out", "both"]),
-            "Always insert diodes on ports with the specified polarities.",
-            default="in",
+            "HEURISTIC_ANTENNA_THRESHOLD",
+            Optional[Decimal],
+            "A manhattan distance above which a diode is recommended to be inserted by a heuristic inserter. If not specified, the heuristic inserter will typically use a default value.",
+            units="µm",
         ),
     ]
 
@@ -318,8 +368,14 @@ class HeuristicDiodeInsertion(OdbpyStep):
                 cell,
                 "--diode-pin",
                 pin,
-                "--port-protect",
-                self.config["DIODE_ON_PORTS"].value,
             ]
             + threshold_opts
         )
+
+    def run(self, **kwargs) -> State:
+        if self.config["GPL_CELL_PADDING"] == 0:
+            warn(
+                "'GPL_CELL_PADDING' is set to 0. This step may cause overlap failures."
+            )
+
+        return super().run(**kwargs)