+++ Changelog.md
@@ -1,4 +1,23 @@
+# 2.0.0-a11
+* Update OpenROAD, Add ABC patch to use system zlib
+* Adds SDC files as an input to `OpenROADStep`, `NetlistSTA` and `LayoutSTA` steps
+* Add `sdc_reader.tcl`: a hook script for reading in SDC files while handling deprecated variables
+* Replace deprecated variables in base.sdc
+    * Properly use TIME_DERATING_CONSTRAINT in base.sdc 
+    * Properly use SYNTH_DRIVING_CELL in base.sdc
+    * Properly use SYNTH_CLK_DRIVING_CELL in base.sdc
+
+# 2.0.0-a10
+* Add `wrapper.tcl` to capture errors in Magic scripts.
+* Fix instances of a deprecated variable was used in Magic scripts.
+
+# 2.0.0-a9
+
+* Add port diode insertion script.
+* Fix formula for calculating `FP_TARGET_DENSITY_PCT`.
+
 # 2.0.0-a8
+
 * Update `volare` dependency.
 * Update `magic` version + make `magic` nix derivation more resilient.